split checks into files
